order operations refined

+++ src/main/kotlin/com/vaddshah/ktjooq/features/orders/dtos/Order.kt
@@ -1,16 +1,16 @@
 package com.vaddshah.ktjooq.features.orders.dtos
 
-import com.vaddshah.ktjooq.features.products.dtos.Product
 import com.vaddshah.ktjooq.features.users.dtos.UserResponse
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class OrderItem(
     val id: Long,
     val createdAt: LocalDateTime,
-    val productId : Long,
+    val productId: Long,
     val productName: String,
     val quantity: Int,
-    val price: Double,
+    val price: BigDecimal,
 )
 
 data class Order(
@@ -18,7 +18,7 @@ data class Order(
     val userId: Long,
     val user: UserResponse? = null,
     val status: String,
-    val total: Double,
+    val total: BigDecimal,
     val items: List<OrderItem>,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime? = null,
